Group frametime, io and misc functions separately
